+++ src/main/scala/models/Entry.scala
@@ -3,4 +3,4 @@ package models
 /**
   * Created by Pedro Lopes on 20/06/2017.
   */
-case class Entry(EntryId: Int, CategoryId: Int, UserId: Int, Name: String, Amount: Double, Date: String)
+case class Entry(EntryId: Long, CategoryId: Long, UserId: Long, Name: String, Amount: Double, Date: String)
