+++ src/main/scala/models/Group.scala
@@ -3,4 +3,4 @@ package models
 /**
   * Created by Pedro Lopes on 20/06/2017.
   */
-case class Group (GroupId: Int, Name: String, Balance: Double = 0)
+case class Group (GroupId: Long, Name: String, Balance: Double = 0)
